移动旧地址到/opt/edpatch 添加判断条件 新地址不存在
修复移动旧txt到/opt/edpatch中时，如果新地址存在则会抛出异常的问题（容器设置 - 渲染设置无法显示）

+++ app/src/main/java/com/example/datainsert/exagear/FAB/dialogfragment/PulseAudio.java
@@ -336,7 +336,7 @@ public class PulseAudio extends BaseFragment {
     public void callWhenFirstStart(AppCompatActivity activity) {
         //移除旧文件
         File oldWorkDir = new File(activity.getFilesDir(),"pulseaudio-xsdl");
-        if(oldWorkDir.exists()) {
+        if(oldWorkDir.exists() && !paWorkDir.exists()) {
             try {
                 FileUtils.moveDirectory(oldWorkDir,paWorkDir);
             } catch (IOException e) {
